+++ test/errors.js
@@ -24,7 +24,7 @@ test.serial('@errors => rejects operations on http error (401)', async (t) => {
   const config = createModuleConfig();
   const oauth2 = oauth2Module.create(config);
 
-  const error = await t.throwsAsync(() => oauth2.authorizationCode.getToken(tokenParams), Error);
+  const error = await t.throwsAsync(() => oauth2.authorizationCode.getToken(tokenParams), { instanceOf: Error });
 
   scope.done();
 
@@ -46,7 +46,7 @@ test.serial('@errors => rejects operations on http error (500)', async (t) => {
   const config = createModuleConfig();
   const oauth2 = oauth2Module.create(config);
 
-  const error = await t.throwsAsync(() => oauth2.authorizationCode.getToken(tokenParams), Error);
+  const error = await t.throwsAsync(() => oauth2.authorizationCode.getToken(tokenParams), { instanceOf: Error });
 
   scope.done();